Alert based on reliability, not single failures
At the moment, this concourse job will alert #govuk-2ndline on slack if
there's a single failure.

The job isn't 100% reliable though, so these failures happen
semi-regularly. This makes the channel noisy, and inevitably leads to
alert blindness.

Instead, I propose that we alert based on the metrics for this pipeline:

https://grafana.monitoring.cd.gds-reliability.engineering/d/1go_x9sMz/concourse-alerts?orgId=1

Big Concourse ships its metrics to Prometheus, which can then be used in
Grafana.

I've set up an alert in Grafana (by hand - no infra-as-code yet) which
will ping slack if the reliability of this job falls below 70% for a
rolling 1 day window.

With this in place, we can remove the notification that fires every time
the job fails.

Alternative option
------------------

Instead of alerting based on metrics, we could configure the
mirror-repos task to retry several times. This would be my preference
ordinarily, but I'm a little concerned about putting a retry on a job
which has a 90 minute timeout and regularly takes > 1 hour to run.