Rate-limit the creation of new repositories
We're hitting a `Aws::CodeCommit::Errors::ThrottlingException`
rate limit.

AWS is not forthcoming with the details of what the actual limit
is, but it does recommending staggering the intervals of the API
calls in the first instance (https://aws.amazon.com/premiumsupport/knowledge-center/sns-throttling-exception-errors/),
so I've added a `sleep` and picked a somewhat arbitrary 5 seconds.