Handle edge case where repo has already been cloned
We've just had a situation where a repo was cloned locally, but
then the script failed at the "Doing some cleanup" stage, meaning
that we had a stray directory hanging around. A subsquent run of
the script failed with:

```
Checking if static is up to date... updating
Fetching static
fatal: destination path 'static.git' already exists and is not an empty directory.
Build step 'Execute shell' marked build as failure
Finished: FAILURE
```

We should handle this by ensuring the local working directory is
clear of the repo first.